Implement CartPoleSwingUpContinuous
